requirements.txt
Flask
gunicorn
langchain-google-genai
langchain-core
assemblyai
python-dotenv 